ユニットテスト考慮して、ファイル構成変更
JavaScriptファイルを、テストフレームワーク適用を考慮して構成変更した。
