+++ client_script/view_battery/brw_jq/ui_panel.js
@@ -1,8 +1,6 @@
 /*
     [ui_panel.js]
-
     encoding=UTF-8
-
 */
