+++ client_script/view_battery/brw_lib/tiny-cookie/tiny-cookie.js
